+++ Source/HtmlRenderer/Entities/RRect.cs
@@ -233,8 +233,8 @@ namespace HtmlRenderer.Entities
         /// </param>
         public static bool operator ==(RRect left, RRect right)
         {
-            if( Math.Abs(left.X - (double)right.X) < 0.001 && Math.Abs(left.Y - (double)right.Y) < 0.001 && Math.Abs(left.Width - (double)right.Width) < 0.001 )
-                return Math.Abs(left.Height - (double)right.Height) < 0.001;
+            if( Math.Abs(left.X - right.X) < 0.001 && Math.Abs(left.Y - right.Y) < 0.001 && Math.Abs(left.Width - right.Width) < 0.001 )
+                return Math.Abs(left.Height - right.Height) < 0.001;
             else
                 return false;
         }
@@ -289,8 +289,8 @@ namespace HtmlRenderer.Entities
             if( !( obj is RRect ) )
                 return false;
             var rectangleF = (RRect)obj;
-            if( Math.Abs(rectangleF.X - (double)X) < 0.001 && Math.Abs(rectangleF.Y - (double)Y) < 0.001 && Math.Abs(rectangleF.Width - (double)Width) < 0.001 )
-                return Math.Abs(rectangleF.Height - (double)Height) < 0.001;
+            if( Math.Abs(rectangleF.X - X) < 0.001 && Math.Abs(rectangleF.Y - Y) < 0.001 && Math.Abs(rectangleF.Width - Width) < 0.001 )
+                return Math.Abs(rectangleF.Height - Height) < 0.001;
             else
                 return false;
         }
@@ -307,8 +307,8 @@ namespace HtmlRenderer.Entities
         /// <param name="y">The y-coordinate of the point to test. </param>
         public bool Contains(double x, double y)
         {
-            if( X <= (double)x && x < X + (double)Width && Y <= (double)y )
-                return y < Y + (double)Height;
+            if( X <= x && x < X + Width && Y <= y )
+                return y < Y + Height;
             else
                 return false;
         }
@@ -342,8 +342,8 @@ namespace HtmlRenderer.Entities
         /// </param>
         public bool Contains(RRect rect)
         {
-            if( X <= (double)rect.X && rect.X + (double)rect.Width <= X + (double)Width && Y <= (double)rect.Y )
-                return rect.Y + (double)rect.Height <= Y + (double)Height;
+            if( X <= rect.X && rect.X + rect.Width <= X + Width && Y <= rect.Y )
+                return rect.Y + rect.Height <= Y + Height;
             else
                 return false;
         }
@@ -425,7 +425,7 @@ namespace HtmlRenderer.Entities
             double num1 = Math.Min(a.X + a.Width, b.X + b.Width);
             double y = Math.Max(a.Y, b.Y);
             double num2 = Math.Min(a.Y + a.Height, b.Y + b.Height);
-            if( num1 >= (double)x && num2 >= (double)y )
+            if( num1 >= x && num2 >= y )
                 return new RRect(x, y, num1 - x, num2 - y);
             else
                 return Empty;
@@ -440,8 +440,8 @@ namespace HtmlRenderer.Entities
         /// <param name="rect">The rectangle to test. </param>
         public bool IntersectsWith(RRect rect)
         {
-            if( rect.X < X + (double)Width && X < rect.X + (double)rect.Width && rect.Y < Y + (double)Height )
-                return Y < rect.Y + (double)rect.Height;
+            if( rect.X < X + Width && X < rect.X + rect.Width && rect.Y < Y + Height )
+                return Y < rect.Y + rect.Height;
             else
                 return false;
         }
